refactor(devframe): split dump runtime into `collect.ts`, keep `dump/index.ts` as barrel
Move `dumpFunctions`, `createClientFromDump`, `getDefinitionsWithDumps`
and their private helpers from `rpc/dump/index.ts` into a new
`rpc/dump/collect.ts`. `index.ts` is now a three-line barrel re-exporting
`./collect`, `./error`, `./static`. Repoint `static.ts` at `./collect` so
the barrel does not reintroduce a `dump/static → dump → dump/static`
cycle.
